:recycle: refactor: 初步完成取色器模型目录结构重构
