+++ packages/color-picker/package.json
@@ -33,7 +33,7 @@
     "reactcss": "^1.2.3",
     "rxjs": "^7.5.5",
     "tinycolor2": "^1.4.2",
-    "zustand": "^3.7.1"
+    "zustand": "^4.0.0"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.181",
