+++ packages/color-picker/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './swatch';
+export * from './component';
+export * from './store';
